split core-worker-process
Revert "split core-worker-process"

This reverts commit 4d5711b3063f62bb0bed6c9b4136deac72737f1b.
