+++ src/ray/core_worker/core_worker.h
@@ -24,6 +24,7 @@
 #include "ray/core_worker/common.h"
 #include "ray/core_worker/context.h"
 #include "ray/core_worker/core_worker_options.h"
+#include "ray/core_worker/core_worker_process.h"
 #include "ray/core_worker/future_resolver.h"
 #include "ray/core_worker/gcs_server_address_updater.h"
 #include "ray/core_worker/lease_policy.h"
@@ -56,160 +57,6 @@
 
 namespace ray {
 namespace core {
-
-class CoreWorker;
-
-/// Lifecycle management of one or more `CoreWorker` instances in a process.
-///
-/// To start a driver in the current process:
-///     CoreWorkerOptions options = {
-///         WorkerType::DRIVER,             // worker_type
-///         ...,                            // other arguments
-///         1,                              // num_workers
-///     };
-///     CoreWorkerProcess::Initialize(options);
-///
-/// To shutdown a driver in the current process:
-///     CoreWorkerProcess::Shutdown();
-///
-/// To start one or more workers in the current process:
-///     CoreWorkerOptions options = {
-///         WorkerType::WORKER,             // worker_type
-///         ...,                            // other arguments
-///         num_workers,                    // num_workers
-///     };
-///     CoreWorkerProcess::Initialize(options);
-///     ...                                 // Do other stuff
-///     CoreWorkerProcess::RunTaskExecutionLoop();
-///
-/// To shutdown a worker in the current process, return a system exit status (with status
-/// code `IntentionalSystemExit` or `UnexpectedSystemExit`) in the task execution
-/// callback.
-///
-/// If more than 1 worker is started, only the threads which invoke the
-/// `task_execution_callback` will be automatically associated with the corresponding
-/// worker. If you started your own threads and you want to use core worker APIs in these
-/// threads, remember to call `CoreWorkerProcess::SetCurrentThreadWorkerId(worker_id)`
-/// once in the new thread before calling core worker APIs, to associate the current
-/// thread with a worker. You can obtain the worker ID via
-/// `CoreWorkerProcess::GetCoreWorker()->GetWorkerID()`. Currently a Java worker process
-/// starts multiple workers by default, but can be configured to start only 1 worker by
-/// speicifying `num_java_workers_per_process` in the job config.
-///
-/// If only 1 worker is started (either because the worker type is driver, or the
-/// `num_workers` in `CoreWorkerOptions` is set to 1), all threads will be automatically
-/// associated to the only worker. Then no need to call `SetCurrentThreadWorkerId` in
-/// your own threads. Currently a Python worker process starts only 1 worker.
-class CoreWorkerProcess {
- public:
-  ///
-  /// Public methods used in both DRIVER and WORKER mode.
-  ///
-
-  /// Initialize core workers at the process level.
-  ///
-  /// \param[in] options The various initialization options.
-  static void Initialize(const CoreWorkerOptions &options);
-
-  /// Get the core worker associated with the current thread.
-  /// NOTE (kfstorm): Here we return a reference instead of a `shared_ptr` to make sure
-  /// `CoreWorkerProcess` has full control of the destruction timing of `CoreWorker`.
-  static CoreWorker &GetCoreWorker();
-
-  /// Try to get the `CoreWorker` instance by worker ID.
-  /// If the current thread is not associated with a core worker, returns a null pointer.
-  ///
-  /// \param[in] workerId The worker ID.
-  /// \return The `CoreWorker` instance.
-  static std::shared_ptr<CoreWorker> TryGetWorker(const WorkerID &worker_id);
-
-  /// Set the core worker associated with the current thread by worker ID.
-  /// Currently used by Java worker only.
-  ///
-  /// \param worker_id The worker ID of the core worker instance.
-  static void SetCurrentThreadWorkerId(const WorkerID &worker_id);
-
-  /// Whether the current process has been initialized for core worker.
-  static bool IsInitialized();
-
-  ///
-  /// Public methods used in DRIVER mode only.
-  ///
-
-  /// Shutdown the driver completely at the process level.
-  static void Shutdown();
-
-  ///
-  /// Public methods used in WORKER mode only.
-  ///
-
-  /// Start receiving and executing tasks.
-  static void RunTaskExecutionLoop();
-
-  // The destructor is not to be used as a public API, but it's required by smart
-  // pointers.
-  ~CoreWorkerProcess();
-
- private:
-  /// Create an `CoreWorkerProcess` with proper options.
-  ///
-  /// \param[in] options The various initialization options.
-  CoreWorkerProcess(const CoreWorkerOptions &options);
-
-  /// Check that the core worker environment is initialized for this process.
-  ///
-  /// \param[in] quick_exit If set to true, quick exit if uninitialized without
-  /// crash.
-  /// \return Void.
-  static void EnsureInitialized(bool quick_exit);
-
-  static void HandleAtExit();
-
-  void InitializeSystemConfig();
-
-  /// Check that if the global worker should be created on construction.
-  bool ShouldCreateGlobalWorkerOnConstruction() const;
-
-  /// Get the `CoreWorker` instance by worker ID.
-  ///
-  /// \param[in] workerId The worker ID.
-  /// \return The `CoreWorker` instance.
-  std::shared_ptr<CoreWorker> GetWorker(const WorkerID &worker_id) const
-      LOCKS_EXCLUDED(mutex_);
-
-  /// Create a new `CoreWorker` instance.
-  ///
-  /// \return The newly created `CoreWorker` instance.
-  std::shared_ptr<CoreWorker> CreateWorker() LOCKS_EXCLUDED(mutex_);
-
-  /// Remove an existing `CoreWorker` instance.
-  ///
-  /// \param[in] The existing `CoreWorker` instance.
-  /// \return Void.
-  void RemoveWorker(std::shared_ptr<CoreWorker> worker) LOCKS_EXCLUDED(mutex_);
-
-  /// Get the `GlobalWorker` instance, if the number of workers is 1.
-  std::shared_ptr<CoreWorker> GetGlobalWorker() LOCKS_EXCLUDED(mutex_);
-
-  /// The various options.
-  const CoreWorkerOptions options_;
-
-  /// The core worker instance associated with the current thread.
-  /// Use weak_ptr here to avoid memory leak due to multi-threading.
-  static thread_local std::weak_ptr<CoreWorker> current_core_worker_;
-
-  /// The only core worker instance, if the number of workers is 1.
-  std::shared_ptr<CoreWorker> global_worker_ GUARDED_BY(mutex_);
-
-  /// The worker ID of the global worker, if the number of workers is 1.
-  const WorkerID global_worker_id_;
-
-  /// Map from worker ID to worker.
-  std::unordered_map<WorkerID, std::shared_ptr<CoreWorker>> workers_ GUARDED_BY(mutex_);
-
-  /// To protect access to workers_ and global_worker_
-  mutable absl::Mutex mutex_;
-};
 
 /// The root class that contains all the core and language-independent functionalities
 /// of the worker. This class is supposed to be used to implement app-language (Java,
